requirements.txt
torch>=1.9.0
matplotlib>=3.7.0
numpy>=1.21.0
Pillow>=8.0.0
